s3_create/requirements.txt
aws-cdk.core
aws-cdk.aws_s3


